fix(api): type AuthService.discover's return as a DiscoverResponse DTO
discover() returned a raw Dict[str, Any] that the router re-validated into
DiscoverResponse; the DTO now lives in core/auth/dtos.py and the service
returns it natively (API-9).
